+++ gazebo/gui/building/LevelInspectorDialog.hh
@@ -55,6 +55,10 @@ namespace gazebo
       /// \return Floor color.
       public: QColor GetFloorColor() const;
 
+      /// \brief Get the texture of this level's floor.
+      /// \return Texture.
+      public: QString GetFloorTexture() const;
+
       /// \brief Set the name of the level.
       /// \param[in] _levelName New level name.
       public: void SetLevelName(const std::string &_levelName);
@@ -66,6 +70,10 @@ namespace gazebo
       /// \brief Set the color of this level's floor.
       /// \param[in] _color Color.
       public: void SetFloorColor(const QColor _color);
+
+      /// \brief Set the texture of this level's floor.
+      /// \param[in] _texture Texture.
+      public: void SetFloorTexture(const QString _texture);
 
       /// \brief Qt signal emitted to indicate that changes should be applied.
       Q_SIGNALS: void Applied();
@@ -91,14 +99,17 @@ namespace gazebo
       /// \brief Spin box for configuring the floor thickness.
       private: QDoubleSpinBox *floorThicknessSpinBox;
 
-      /// \brief Combo box for configuring the floor material.
-      private: QComboBox *materialComboBox;
-
       /// \brief Combo box for selecting the color of the floor.
       private: QComboBox *floorColorComboBox;
 
       /// \brief Vector of color options for the floor.
       private: std::vector<QColor> floorColorList;
+
+      /// \brief Combo box for selecting the texture of the floor.
+      private: QComboBox *floorTextureComboBox;
+
+      /// \brief Vector of texture options.
+      private: std::vector<QString> floorTextureList;
     };
     /// \}
   }