+++ Changelog.md
@@ -6,13 +6,13 @@
     * [Pull request #1278](https://bitbucket.org/osrf/gazebo/pull-request/1278)
     * [Issue #1327](https://bitbucket.org/osrf/gazebo/issue/1327)
 
-1. Added a GUI timer plugin that facilitates the display and control a timer inside the Gazebo UI. 
+1. Added a GUI timer plugin that facilitates the display and control a timer inside the Gazebo UI.
     * [Pull request #1270](https://bitbucket.org/osrf/gazebo/pull-request/1270)
 
-1. Added ability to load plugins via SDF. 
+1. Added ability to load plugins via SDF.
     * [Pull request #1261](https://bitbucket.org/osrf/gazebo/pull-request/1261)
 
-1. Added GUIEvent to hide/show the left GUI pane. 
+1. Added GUIEvent to hide/show the left GUI pane.
     * [Pull request #1269](https://bitbucket.org/osrf/gazebo/pull-request/1269)
 
 1. Modified KeyEventHandler and GLWidget so that hotkeys can be suppressed by custom KeyEvents set up by developers
@@ -65,6 +65,12 @@
 
 1. Added GLWidget::OnModelEditor when model editor is triggered, and MainWindow::OnEditorGroup to manually uncheck editor actions.
     * [Pull request #1283](https://bitbucket.org/osrf/gazebo/pull-request/1283)
+
+1. Added Collision, Geometry, Inertial, Surface Msg-to-SDF conversion functions.
+    * [Pull request #1315](https://bitbucket.org/osrf/gazebo/pull-request/1315)
+
+1. Added "button modifier" fields (control, shift, and alt) to common::KeyEvent.
+    * [Pull request #1325](https://bitbucket.org/osrf/gazebo/pull-request/1325)
 
 1. Building editor updates
     1. Fixed inspector resizing.
@@ -134,9 +140,13 @@
         * [Pull request #1302](https://bitbucket.org/osrf/gazebo/pull-request/1302)
         * [Issue #1323](https://bitbucket.org/osrf/gazebo/issue/1323)
 
+    1. Added texture picker to inspector dialogs.
+        * [Pull request #1306](https://bitbucket.org/osrf/gazebo/pull-request/1306)
+
 1. Model editor updates
     1. Fix adding/removing event filters .
         * [Pull request #1279](https://bitbucket.org/osrf/gazebo/pull-request/1279)
+
 
 ## Gazebo 4.0